requirements.txt
python-dotenv
openai>=1.40.0
langchain
langchain-openai
langchain-chroma
langchain_community
chromadb
pypdf
tiktoken
unstructured
streamlit
streamlit-authenticator
pyyaml
jinja2